sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insight-profiler"
version = "0.1.0"
description = "Python SDK for the Insight Profiler — high-precision power measurement tool by EmbedLab-Tech"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "EmbedLab-Tech"},
]
keywords = ["power-profiler", "embedded", "iot", "energy", "measurement"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Embedded Systems",
]
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
    "mypy",
]
plot = [
    "matplotlib>=3.8",
    "numpy>=1.26",
]

[project.urls]
Homepage = "https://github.com/EmbedLab-Tech/insight-profiler"
Issues = "https://github.com/EmbedLab-Tech/insight-profiler/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
